requirements.txt
websockets>=12.0
aiohttp>=3.9.0
aiofiles>=23.0.0
python-dotenv>=1.0.0